requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
Beaker==1.10.0
cairocffi==1.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
dasbus==1.4
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
Django==3.2.6
evdev==1.1.2
filelock==3.0.12
idna==2.10
langtable==0.0.56
lutris==0.5.8.4
lxml==4.6.3
Mako==1.1.4.dev0
MarkupSafe==1.1.1
nftables==0.1
olefile==0.46
packaging==20.9
Paste==3.5.0
pexpect==4.8.0
phonenumbers==8.12.28
pid==2.2.3
Pillow==8.1.2
pipenv==2021.5.29
ply==3.11
productmd==1.33
ptyprocess==0.6.0
pwquality==1.4.4
pycairo==1.20.0
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.1
PyGObject==3.40.1
pykickstart==3.32
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.7
PySocks==1.7.1
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python-meh==0.50
pytz==2021.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
qtile==0.14.2
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.16.1.3
selinux==3.2
sepolicy==3.2
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.8
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.1
sqlparse==0.4.1
systemd-python==234
Tempita==0.5.1
toml==0.10.2
urllib3==1.25.10
virtualenv==20.4.0
virtualenv-clone==0.5.6
xcffib==0.9.0
